Put RemoteMessagingConfigMatcherProvider in a separate file
